pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "startup-radar"
dynamic = ["version"]
description = "Personal startup discovery radar with Streamlit dashboard"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.2",
    "pandas>=2.0.0",
    "streamlit>=1.30.0",
    "duckduckgo-search>=6.0.0",
    "python-docx>=1.1.0",
    "typer>=0.12",
    "pydantic>=2.5",
]

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
]

[project.scripts]
startup-radar = "startup_radar.cli:app"

[tool.setuptools_scm]
fallback_version = "0.1.0"
# Existing tags are phase-0..3 (not semver); scm will emit dev-style versions
# like "0.1.0.dev{N}+g{sha}" from HEAD. Fallback keeps source-tarball builds sane.

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "vcrpy>=6.0",
    "pytest-recording>=0.13.4",
]

[tool.setuptools]
packages = [
    "startup_radar",
    "startup_radar.sources",
    "startup_radar.parsing",
    "startup_radar.research",
    "startup_radar.config",
]
py-modules = [
    "app",
    "database",
    "connections",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [".venv", "build", "dist", "reports", "logs"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = [
    "E501",   # line length — let formatter handle
    "E722",   # bare except — Phase 4 will replace with structured logging
    "B904",   # raise from — same
    "B007",   # unused loop var — common in pandas iterrows
    "UP006",  # PEP585 — keep `list[T]` style consistency
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "B011"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
files = [
    "startup_radar/models.py",
    "startup_radar/parsing",
    "startup_radar/config",
    "startup_radar/filters.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --strict-markers --cov=startup_radar --cov-report=term-missing --cov-report=xml"
markers = [
    "integration: requires network or external services",
    "vcr: cassette-backed test (see tests/fixtures/cassettes/)",
]

[tool.coverage.run]
source = ["startup_radar"]
omit = [
    "startup_radar/cli.py",
    "startup_radar/research/deepdive.py",
]

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
